+++ CHANGELOG.md
@@ -2,6 +2,9 @@
 
 ## Master Dev Branch
 
+### 2026-09-02
+- Add Twig cache for compiled templates
+
 ### 2026-09-01
 - OIDC OPENIDC_REQUIRE_VERIFIED_EMAIL, OPENIDC_DEBUG_CLAIMS, OPENIDC_PREFERRED_USERNAME_ATTR
 - Don't allow admin account takeover by external auth